\name{authenticateWebServiceCall}
\alias{authenticateWebServiceCall}

\title{
Authenticate with the rest webservice
}
\description{
THis function autheticate with the REST webservice and in turn returns a curl handle for subsequent calls
}
\usage{
authenticateWebServiceCall(username, password)
}

\arguments{
  \item{username}{
Login username
}
  \item{password}{
Login password
}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
Prasanna_Teli@infosys.com
}
\note{

}



\seealso{

}
\examples{
# authenticateWebServiceCall(username,password)
}