\name{createGetFileLocationJson}
\alias{createGetFileLocationJson}

\title{
Create the json object for GetFilePath web service
}
\description{
This function creates the json object for IIP Webservice GetFilePath
}
\usage{
createGetFileLocationJson(tableName, dataSource, workspaceName)
}
\arguments{
  \item{tableName}{
  Name of the IIP table which contains the ingested data 

}
  \item{dataSource}{
  Name of the IIP data source

}
  \item{workspaceName}{
  Name of the IIP workspace

}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}



\seealso{

}
\examples{
#createGetFileLocationJson(tableName="tableName", dataSource="dataSource", workspaceName="workspaceName")

}