\name{createLocalDirectory}
\alias{createLocalDirectory}
\title{
create a local directory
}
\description{
create a directory with dirName at given parent directory location
}
\usage{
createLocalDirectory(parentDir, dirName)
}
\arguments{
  \item{parentDir}{
  Location of the Parent directory where the directory is to be created

}
  \item{dirName}{
  Name of the directory

}
}
\details{
}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}



\seealso{

}
\examples{

#createLocalDirectory("/home","/test")
}
