\name{deleteLocalDirectory}
\alias{deleteLocalDirectory}

\title{
delete the local directory
}
\description{
delete the given local directory from parent directory
}
\usage{
deleteLocalDirectory(dirName)
}

\arguments{
  \item{dirName}{
  Name of the local directory which has to be deleted

}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}



\seealso{
}
\examples{
deleteLocalDirectory("/test")

}
