\name{get_console_pwd}
\alias{get_console_pwd}

\title{
Get the password from user
}
\description{
This function allows to get the masked password from user on console.
}
\usage{
get_console_pwd()
}
\details{

}
\value{

}
\references{

}
\author{
Prasanna_Teli@infosys.com
}
\note{

}

\seealso{

}
\examples{
#get_console_pwd()
}
