\name{get_ui_pass}
\alias{get_ui_pass}

\title{
Input Password
}
\description{
Function to enter the password in a masked format by using a Windows Pop-up
}
\usage{
get_ui_pass()
}

\details{

}
\value{

}
\references{

}
\author{
Prasanna_Teli@infosys.com
}
\note{

}



\seealso{

}
\examples{
#get_ui_pass()
}
