\name{getExactFilePath}
\alias{getExactFilePath}
\title{
get the absolute file location 
}
\description{
get the file path given the directory location
}
\usage{
getExactFilePath(dirLocation)
}

\arguments{
  \item{dirLocation}{
Location of the directory for which the absolute path has to be determined.
}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}



\seealso{

}
\examples{

#getExactFilePath("FilePathLocation")


}

