\name{getPasswordFromUser}
\alias{getPasswordFromUser}

\title{
Input Password
}
\description{
Function to enter the password in a masked format by either using a Windows Pop-up or console
}
\usage{
getPasswordFromUser()
}
\details{

}
\value{

}
\references{

}
\author{
poorva_agrawal@infosys.com
}
\note{

}



\seealso{

}
\examples{
#getPasswordFromUser()
}
