\name{IIP.decrypt}
\alias{IIP.decrypt}

\title{
Decrypt password
}
\description{
Utility to decrypt the password which the user has entered and is stored in encrypted format using key(RSA 2048bits).
}
\usage{
IIP.decrypt(encryptedPwdFile = NULL, privateKeyFile = NULL)
}

\arguments{
  \item{encryptedPwdFile}{
Location of the Encrypted Password File.
}
  \item{privateKeyFile}{
Location of the Private Key file which stores the RSA Private Key.
}
}
\details{

}
\value{

}
\references{

}
\author{
Prasanna_Teli@infosys.com
}
\note{

}


\seealso{

}
\examples{
#IIP.decrypt( encryptedPwdFile = "/home/test")

}
