\name{IIP.encrypt}
\alias{IIP.encrypt}

\title{
Encrypt passwords
}
\description{
Function to encrypt passwords using IIP jar
}
\usage{
IIP.encrypt(password)
}

\arguments{
  \item{password}{
Plaintext Original Password
}
}
\details{

}
\value{

}
\references{

}
\author{
poorva_agrawal@infosys.com, Prasanna_Teli@infosys.com
}
\note{

}

\seealso{

}
\examples{
#IIP.encrypt(password)
}

