\name{IIP.JDBCConnection}
\alias{IIP.JDBCConnection}
\title{
Functionality to get IIP views using JDBC connection
}
\description{
Functionality to get IIP views using JDBC connection
}
\usage{
IIP.JDBCConnection(datasource = "iip")
}

\arguments{
  \item{datasource}{
Specify the data source to pull the data from. 
Takes in the type as "iip", "hive", "mysql", "postgres"
}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
}
\note{

}


\seealso{

}
\examples{
#IIP.JDBCConnection("iip")
}