\name{IIPR-package}
\alias{IIPR-package}
\alias{IIPR}
\docType{package}
\title{
Package containing utilities for IIP
}
\description{
the IIPR package has been built to allow data scientists and analysts access views created in IIP
}
\details{
\tabular{ll}{
Package: \tab IIPR\cr
Type: \tab Package\cr
Version: \tab 2.3\cr
Date: \tab 2017-01-25\cr
License: \tab Apache License, Version 2.0\cr
}

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\references{

}
\keyword{ package }
\seealso{

}
\examples{

}
