\name{uploadToHDFS}
\alias{uploadToHDFS}
\title{
uploadToHDFS
}
\description{
This program uploads the file from local system to HDFS
}
\usage{
uploadToHDFS(inputFilePath, hdfsFilePath)
}

\arguments{
  \item{inputFilePath}{
the path of local file to be uploaded
}
  \item{hdfsFilePath}{
hdfs path for uploading the file
}
}
\details{

}
\value{
}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}


\examples{

#uploadToHDFS(inputFilePath="LocalFilePathLocation", hdfsFilePath="HDFSFilePathLocation")

}