\name{getUserRole}
\alias{getUserRole}

\title{
Fetch the user roles and id
}
\description{
This function fetches all the roles and id available for the user in IIP
}
\usage{
getUserRole(username, password, curl)
}

\arguments{
  \item{username}{
Login username
}
  \item{password}{
Login password
}
  \item{curl}{
The curl handle, which is returned from authenticateWebServiceCall()
}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}


\seealso{

}
\examples{

#getUserRole(username, password, curl)

}
