\name{IIP.decrypt}
\alias{IIP.decrypt}

\title{
Decrypt password
}
\description{
Utility to decrypt the encrypted password using IIP jar
}
\usage{
IIP.decrypt(encryptedPWD)
}

\arguments{
  \item{encryptedPWD}{
The encrypted password
}
}
\details{

}
\value{

}
\references{

}
\author{
poorva_agrawal@infosys.com, Prasanna_Teli@infosys.com
}
\note{

}


\seealso{

}
\examples{
#IIP.decrypt(encryptedPwd)

}
