\name{IIP.decrypt}
\alias{IIP.decrypt}

\title{
Decrypt password
}
\description{
Utility to decrypt the encrypted password using IIP jar
}
\usage{
IIP.decrypt(encryptedPwdFile, privateKeyFile)
}

\arguments{
  \item{encryptedPwdFile}{absolute path of encrypted password file}
  \item{privateKeyFile}{absolute path of private key file}

}
\details{

}
\value{

}
\references{

}
\author{
poorva_agrawal@infosys.com, Prasanna_Teli@infosys.com
}
\note{

}


\seealso{

}
\examples{
#IIP.decrypt(encryptedPwdFile, privateKeyFile)

}
