\name{IIP.encrypt}
\alias{IIP.encrypt}

\title{
Encrypt passwords
}
\description{
Function to encrypt passwords using IIP jar
}
\usage{
IIP.encrypt(pwd = NULL, pwdFile = NULL, encryptedFilePath = NULL)
}

\arguments{
  \item{pwd}{Plaintext Original Password}
  \item{pwdFile}{PlainText password file}
  \item{encryptedFilePath}{Excrypted password file}
}
\details{

}
\value{

}
\references{

}
\author{
poorva_agrawal@infosys.com, Prasanna_Teli@infosys.com
}
\note{

}

\seealso{

}
\examples{
#IIP.encrypt(pwd = password)
}

