\name{IIP.getTableFilePath}
\alias{IIP.getTableFilePath}

\title{
Get the hdfs location of IIP table
}
\description{
Get the file path of IIP table using IIP webservice
}
\usage{
IIP.getTableFilePath(tableName, dataSource, workspaceName)
}

\arguments{
  \item{tableName}{
Name of the IIP table which contains the ingested data 
}
  \item{dataSource}{
Name of the IIP data source
}
  \item{workspaceName}{
Name of the IIP workspace
}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}



\seealso{

}
\examples{
#IIP.getTableFilePath(tableName="TableName", dataSource="DataSourceName", workspaceName="WorkspaceName")
}
